build: enable offline/Linux local build+test via system packages
Make a local CMake build and the unit tests runnable without network access
and on non-MSVC toolchains, while preserving the canonical Windows/MSVC path.

Build configuration:
- Dependencies.cmake now prefers installed packages (find_package for spdlog /
  nlohmann_json / GTest, pkg-config for libssh2) and only falls back to the
  pinned FetchContent sources when they are absent — so a build no longer
  requires cloning from GitHub when system packages exist.
- core/CMakeLists only adds the libssh2 FetchContent include dir when that
  source tree is actually present (a system package carries includes itself).
- CMakePresets gains additive linux-debug configure/build and linux-unit test
  presets that use the system compiler + packages. Windows presets unchanged.

Portability fixes surfaced by GCC -Werror (these only differ from MSVC; no
behaviour change):
- qsizetype index variables no longer truncate to int across the cmdtlm/infra/
  plugin viewmodels, protocol rule support, and several UI views; QRegular
  expression match positions are cast explicitly for setFormat.
- LogViewerView.h was missing #include <QComboBox> (compiled on MSVC only by
  luck of transitive includes).
- An unused helper in PluginView is marked [[maybe_unused]].

Verified on Linux/GCC 13:
  cmake --preset linux-debug
  cmake --build --preset linux-debug
  ctest --preset linux-unit   ->  123/123 passed

+++ src/ui/widgets/CmdTlmHighlighter.cpp
@@ -93,7 +93,8 @@ void CmdTlmHighlighter::highlightBlock(const QString& text)
         auto it = rule.pattern.globalMatch(text);
         while (it.hasNext()) {
             const auto m = it.next();
-            setFormat(m.capturedStart(), m.capturedLength(), rule.format);
+            setFormat(static_cast<int>(m.capturedStart()),
+                      static_cast<int>(m.capturedLength()), rule.format);
         }
     }
 }
